+++ mipidsi/src/lib.rs
@@ -98,7 +98,11 @@ use models::Model;
 mod graphics;
 
 mod test_image;
+use orientation::MemoryMapping;
 pub use test_image::TestImage;
+
+mod orientation;
+pub use orientation::{Rotation, Orientation, InvalidAngleError};
 
 #[cfg(feature = "batch")]
 mod batch;
@@ -245,7 +249,18 @@ where
     // Sets the address window for the display.
     fn set_address_window(&mut self, sx: u16, sy: u16, ex: u16, ey: u16) -> Result<(), Error> {
         // add clipping offsets if present
-        let offset = self.options.window_offset();
+        let mut offset = self.options.display_offset;
+        let mapping = MemoryMapping::from(self.options.orientation);
+        if mapping.reverse_columns {
+            offset.0 = M::FRAMEBUFFER_SIZE.0 - (self.options.display_size.0 + offset.0);
+        }
+        if mapping.reverse_rows {
+            offset.1 = M::FRAMEBUFFER_SIZE.1 - (self.options.display_size.1 + offset.1);
+        }
+        if mapping.swap_rows_and_columns {
+            offset = (offset.1, offset.0);
+        }
+
         let (sx, sy, ex, ey) = (sx + offset.0, sy + offset.1, ex + offset.0, ey + offset.1);
 
         self.dcs.write_command(dcs::SetColumnAddress::new(sx, ex))?;