+++ mipidsi/src/models/st7735s.rs
@@ -8,7 +8,7 @@ use crate::{
         SetPixelFormat, SoftReset, WriteMemoryStart,
     },
     error::InitError,
-    Builder, ColorInversion, Error, ModelOptions,
+    Builder, Error, ModelOptions,
 };
 
 use super::{Dcs, Model};
@@ -18,6 +18,7 @@ pub struct ST7735s;
 
 impl Model for ST7735s {
     type ColorFormat = Rgb565;
+    const FRAMEBUFFER_SIZE: (u16, u16) = (132, 162);
 
     fn init<RST, DELAY, DI>(
         &mut self,
@@ -91,10 +92,7 @@ impl Model for ST7735s {
     }
 
     fn default_options() -> ModelOptions {
-        let mut options = ModelOptions::with_sizes((80, 160), (132, 162));
-        options.set_invert_colors(ColorInversion::Inverted);
-
-        options
+        ModelOptions::full_size::<Self>()
     }
 }