+++ mipidsi/src/options.rs
@@ -1,5 +1,7 @@
 //! [ModelOptions] and other helper types.
 
+use crate::{models::Model, Orientation};
+
 /// [ModelOptions] holds the settings for [Model](crate::Model)s.
 ///
 /// `display_size` being set is the minimum requirement.
@@ -13,45 +15,34 @@ pub struct ModelOptions {
     pub(crate) invert_colors: ColorInversion,
     /// Display refresh order
     pub(crate) refresh_order: RefreshOrder,
-    /// Offset override function returning (w, h) offset for current
-    /// display orientation if display is "clipped" and needs an offset for (e.g. Pico v1)
-    pub(crate) window_offset_handler: fn(&ModelOptions) -> (u16, u16),
     /// Display size (w, h) for given display/model
     pub(crate) display_size: (u16, u16),
-    /// Framebuffer size (w, h) for given display/model
-    pub(crate) framebuffer_size: (u16, u16),
+    /// Display offset (x, y) for given display/model
+    pub(crate) display_offset: (u16, u16),
 }
 
 impl ModelOptions {
-    /// Creates model options for the given display and framebuffer sizes.
-    ///
-    /// All other settings are initialized to their default value.
-    pub fn with_sizes(display_size: (u16, u16), framebuffer_size: (u16, u16)) -> Self {
+    /// Creates model options for the entire framebuffer.
+    pub fn full_size<M: Model>() -> Self {
         Self {
             color_order: ColorOrder::default(),
             orientation: Orientation::default(),
             invert_colors: ColorInversion::default(),
             refresh_order: RefreshOrder::default(),
-            window_offset_handler: no_offset,
-            display_size,
-            framebuffer_size,
+            display_size: M::FRAMEBUFFER_SIZE,
+            display_offset: (0, 0),
         }
     }
 
-    /// Creates model options for the given sizes and offset handler.
-    pub fn with_all(
-        display_size: (u16, u16),
-        framebuffer_size: (u16, u16),
-        window_offset_handler: fn(&ModelOptions) -> (u16, u16),
-    ) -> Self {
+    /// Creates model options for the given size and offset.
+    pub fn with_all(display_size: (u16, u16), display_offset: (u16, u16)) -> Self {
         Self {
             color_order: ColorOrder::default(),
             orientation: Orientation::default(),
             invert_colors: ColorInversion::default(),
             refresh_order: RefreshOrder::default(),
-            window_offset_handler,
             display_size,
-            framebuffer_size,
+            display_offset,
         }
     }
 
@@ -64,36 +55,11 @@ impl ModelOptions {
     ///
     /// Used by models.
     pub(crate) fn display_size(&self) -> (u16, u16) {
-        Self::orient_size(self.display_size, self.orientation())
-    }
-
-    /// Returns framebuffer size based on current orientation and display options.
-    ///
-    /// Used by models. Uses display_size if framebuffer_size is not set.
-    pub(crate) fn framebuffer_size(&self) -> (u16, u16) {
-        let size = if self.framebuffer_size == (0, 0) {
+        if self.orientation.rotation.is_horizontal() {
             self.display_size
         } else {
-            self.framebuffer_size
-        };
-
-        Self::orient_size(size, self.orientation())
-    }
-
-    /// Returns the larger of framebuffer width or height.
-    ///
-    /// Used for scroll area setups.
-    pub(crate) fn framebuffer_size_max(&self) -> u16 {
-        let (w, h) = self.framebuffer_size();
-
-        w.max(h)
-    }
-
-    /// Returns window offset (x, y) based on current orientation and display options.
-    ///
-    /// Used by [Display::set_address_window](crate::Display::set_address_window).
-    pub(crate) fn window_offset(&mut self) -> (u16, u16) {
-        (self.window_offset_handler)(self)
+            (self.display_size.1, self.display_size.0)
+        }
     }
 
     /// Returns the current orientation.
@@ -104,60 +70,6 @@ impl ModelOptions {
     /// Sets the orientation.
     pub fn set_orientation(&mut self, orientation: Orientation) {
         self.orientation = orientation;
-    }
-
-    // Flip size according to orientation, in general
-    fn orient_size(size: (u16, u16), orientation: Orientation) -> (u16, u16) {
-        match orientation {
-            Orientation::Portrait(_) | Orientation::PortraitInverted(_) => size,
-            Orientation::Landscape(_) | Orientation::LandscapeInverted(_) => (size.1, size.0),
-        }
-    }
-}
-
-///
-/// `no_offset` is the default offset provider. It results to 0, 0 in case display_size is == framebuffer_size
-/// and to framebuffer_size - display_size otherwise.
-///
-fn no_offset(options: &ModelOptions) -> (u16, u16) {
-    // do FB size - Display size offset for inverted setups
-    match options.orientation {
-        Orientation::PortraitInverted(_) | Orientation::LandscapeInverted(_) => {
-            let hdiff = options.framebuffer_size.1 - options.display_size.1;
-
-            let mut x = 0;
-            let mut y = 0;
-
-            match options.orientation {
-                Orientation::PortraitInverted(_) => y = hdiff,
-                Orientation::LandscapeInverted(_) => x = hdiff,
-                _ => {}
-            }
-
-            (x, y)
-        }
-        _ => (0, 0),
-    }
-}
-
-///
-/// Display orientation.
-///
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub enum Orientation {
-    /// Portrait orientation, with mirror image parameter
-    Portrait(bool),
-    /// Landscape orientation, with mirror image parameter
-    Landscape(bool),
-    /// Inverted Portrait orientation, with mirror image parameter
-    PortraitInverted(bool),
-    /// Inverted Lanscape orientation, with mirror image parameter
-    LandscapeInverted(bool),
-}
-
-impl Default for Orientation {
-    fn default() -> Self {
-        Self::Portrait(false)
     }
 }